removing direct resource invocations in favor of trying from an agent/CLI
