+++ apps/scan/src/app/(app)/_components/resources/header/method.tsx
