+++ apps/scan/src/app/(app)/_components/resources/utils.ts
