+++ apps/scan/src/app/(app)/admin/tags/_components/resource-executor-modal.tsx
@@ -7,10 +7,10 @@ import {
   DialogTitle,
 } from '@/components/ui/dialog';
 import { api } from '@/trpc/client';
-import { ResourceExecutor } from '@/app/(app)/_components/resources/executor';
+import { ResourceCard } from '@/app/(app)/_components/resources/resource-card';
 import { parseX402Response } from '@/lib/x402';
 import { Loader2 } from 'lucide-react';
-import { getBazaarMethod } from '@/app/(app)/_components/resources/executor/utils';
+import { getBazaarMethod } from '@/app/(app)/_components/resources/utils';
 import {
   Accordion,
   AccordionContent,
@@ -95,17 +95,16 @@ export function ResourceExecutorModal({
             {resource.resource}
           </DialogTitle>
         </DialogHeader>
-        <Accordion type="multiple" defaultValue={[resource.id]}>
-          <ResourceExecutor
-            resource={resource}
-            tags={resource.tags.map(tag => tag.tag)}
-            bazaarMethod={getBazaarMethod(
-              resource.accepts.find(accept => accept.outputSchema)?.outputSchema
-            )}
-            response={parsedResponse.data}
-            className="bg-transparent"
-          />
-        </Accordion>
+        <ResourceCard
+          resource={resource}
+          tags={resource.tags.map(tag => tag.tag)}
+          bazaarMethod={getBazaarMethod(
+            resource.accepts.find(accept => accept.outputSchema)?.outputSchema
+          )}
+          response={parsedResponse.data}
+          className="bg-transparent"
+          isFlat
+        />
         {outputSchema && (
           <Accordion type="single" collapsible className="mt-4">
             <AccordionItem value="output-schema">
